Added print number

+++ 0x06-pointers_arrays_strings/101-print_number.c
@@ -6,7 +6,7 @@
 */
 void print_number(int n)
 {
-long int num = n;
+unsigned int num = n;
 
 if (n < 0)
 {
